math: Fix UB in frexpf
$ math/test-float-frexp
testing float (without inline functions)
UBSAN: Undefined behaviour in ../sysdeps/ieee754/flt-32/s_frexpf.c:36:31 left shift of 4294967295 by 23 cannot be represented in type 'int'
Aborted (core dumped)

+++ sysdeps/ieee754/flt-32/s_frexpf.c
@@ -33,7 +33,7 @@ __frexpf (float x, int *eptr)
     {
       int e = ex - EXPONENT_BIAS + 1;
       *eptr = e;
-      return asfloat (hx - (e << MANTISSA_WIDTH));
+      return asfloat (hx - ((uint32_t) e << MANTISSA_WIDTH));
     }
 
   /* Handle zero, infinity, and NaN.  */
